+++ CHANGELOG.md
@@ -24,9 +24,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Metal ligand: exclude amides and some amines.
 - The Pi stacking interactions have been changed for a more accurate implementation
   (PR #97).
+- The `pdbqt_supplier` will not add explicit hydrogen atoms anymore, to avoid detecting
+  hydrogen bonds with "random" hydrogens that weren't in the PDBQT file.
+- When using the `pdbqt_supplier`, irrelevant warnings and logs have been disabled.
+- Updated the minimal RDKit version to `2021.03.1` 
   
 ### Fixed
 - Dead link in the quickstart notebook for the MDAnalysis quickstart (PR #75, @radifar).
+- The `pdbqt_supplier` now correctly preserves hydrogens from the input PDBQT file.
 
 ## [1.0.0] - 2022-06-07
 ### Added